Possible Soundtrack for game??
This is the first track from the film Dunkirk, composed by Hans Zimmer. It has a motif of a clock ticking and a melancholic, cinematic vibe. Although the track is copyrighted, we are not using this game to make money and are in an educational setting, so we are good.